main.py simplified
